feat(TrailConditionReportView): add TrailConditionReportView and refactored RoadConditionModal into a component
