+++ src/router/index.js
@@ -4,6 +4,7 @@ import FrontIndexView from '@/views/front/FrontIndexView.vue'
 import TrailsIntroView from '@/views/front/TrailsIntroView.vue'
 import TrailsListView from '@/views/front/TrailsListView.vue'
 import TrailInfoView from '@/views/front/TrailInfoView.vue'
+import TrailConditionReportView from '@/views/front/TrailConditionReportView.vue'
 
 import LoginView from '@/views/account/LoginView.vue'
 import RegisterView from '@/views/account/RegisterView.vue'
@@ -37,6 +38,11 @@ const routes = [
         path: 'trail-info/:trail',
         name: 'TrailInfo',
         component: TrailInfoView
+      },
+      {
+        path: 'trail-condition-report',
+        name: 'TrailConditionReport',
+        component: TrailConditionReportView
       }
     ]
   },
